fix: 승인상태 필드 이동 session -> student (4단계)

+++ src/main/java/nextstep/courses/domain/StudentRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface StudentRepository {
-    int save(Student student);
+    long save(Student student);
 
     Optional<Student> findById(Long id);
